[move] tests package changed
